adapting point cloud frmt to int16_t*

+++ src/region.cpp
@@ -1,6 +1,7 @@
 #include <vector>
 
 #include "outliers.h"
+#include "ply.h"
 #include "proposal.h"
 #include "region.h"
 #include "svd.h"
@@ -26,9 +27,9 @@ std::vector<Point> region::segment(std::vector<Point>& points)
 
 std::pair<Point, Point> region::queryBoundary(std::vector<Point>& points)
 {
-    std::vector<float> X(points.size());
-    std::vector<float> Y(points.size());
-    std::vector<float> Z(points.size());
+    std::vector<int16_t> X(points.size());
+    std::vector<int16_t> Y(points.size());
+    std::vector<int16_t> Z(points.size());
     for (auto& point : points) {
         X.push_back(point.m_xyz[0]);
         Y.push_back(point.m_xyz[1]);
@@ -36,12 +37,12 @@ std::pair<Point, Point> region::queryBoundary(std::vector<Point>& points)
     }
 
     /** find boundary given a set of points */
-    float xMax = *std::max_element(X.begin(), X.end());
-    float xMin = *std::min_element(X.begin(), X.end());
-    float yMax = *std::max_element(Y.begin(), Y.end());
-    float yMin = *std::min_element(Y.begin(), Y.end());
-    float zMax = *std::max_element(Z.begin(), Z.end());
-    float zMin = *std::min_element(Z.begin(), Z.end());
+    int16_t xMax = (int16_t)*std::max_element(X.begin(), X.end());
+    int16_t xMin = (int16_t)*std::min_element(X.begin(), X.end());
+    int16_t yMax = (int16_t)*std::max_element(Y.begin(), Y.end());
+    int16_t yMin = (int16_t)*std::min_element(Y.begin(), Y.end());
+    int16_t zMax = (int16_t)*std::max_element(Z.begin(), Z.end());
+    int16_t zMin = (int16_t)*std::min_element(Z.begin(), Z.end());
 
     Point min(xMin, yMin, zMin);
     Point max(xMax, yMax, zMax);
